refactor(core): `Svg` move to legacy
