+++ projects/core/components/label/label.style.less
@@ -13,7 +13,7 @@
         font: var(--tui-font-text-m);
     }
 
-    //TODO: Remove after all inputs migrate to new textfield
+    //TODO: Remove after legacy inputs are dropped
     &:has(tui-primitive-textfield),
     &:has(tui-textarea) {
         flex-direction: column !important;
