+++ projects/core/components/root/root.component.ts
@@ -17,7 +17,6 @@ import {
 import {TuiAlertsComponent} from '@taiga-ui/core/components/alert';
 import {TUI_DIALOGS, TuiDialogsComponent} from '@taiga-ui/core/components/dialog';
 import {TuiScrollControlsComponent} from '@taiga-ui/core/components/scrollbar';
-import {TuiSvgDefsHostComponent} from '@taiga-ui/core/components/svg-defs-host';
 import {TuiDropdownsComponent} from '@taiga-ui/core/directives';
 import {TuiHintsComponent} from '@taiga-ui/core/directives/hint';
 import {TuiBreakpointService} from '@taiga-ui/core/services';
@@ -32,7 +31,6 @@ import {debounceTime, map, of} from 'rxjs';
     imports: [
         NgIf,
         AsyncPipe,
-        TuiSvgDefsHostComponent,
         TuiDropdownsComponent,
         TuiDialogsComponent,
         TuiAlertsComponent,
