+++ projects/core/components/root/root.template.html
@@ -2,7 +2,6 @@
     *ngIf="scrollbars$ | async"
     class="t-root-scrollbar"
 />
-<tui-svg-defs-host />
 <tui-dropdowns>
     <div class="t-root-content">
         <ng-content />
