--- projects/core/components/svg-defs-host/index.ts
@@ -1,1 +1,0 @@
-export * from './svg-defs-host.component';
