--- projects/core/components/svg-defs-host/svg-defs-host.style.less
@@ -1,3 +1,0 @@
-:host {
-    display: none;
-}
