--- projects/core/components/svg-defs-host/svg-defs-host.template.html
@@ -1,18 +1,0 @@
-<svg
-    version="1.1"
-    xmlns:xlink="http://www.w3.org/1999/xlink"
-    xmlns="http://www.w3.org/2000/svg"
->
-    <ng-container *ngIf="isBrowser; else server">
-        <defs
-            *ngFor="let def of items"
-            [innerHTML]="def"
-        />
-    </ng-container>
-    <ng-template #server>
-        <!--svg[outerHTML] - to support SSR, because there is no support of innerHTML for for defs element in domino-->
-        <defs *ngFor="let def of items">
-            <svg [outerHTML]="def" />
-        </defs>
-    </ng-template>
-</svg>
