+++ projects/core/services/index.ts
@@ -2,5 +2,4 @@ export * from './breakpoint.service';
 export * from './dark-theme.service';
 export * from './format-date.service';
 export * from './position.service';
-export * from './svg.service';
 export * from './visual-viewport.service';
