+++ projects/core/tokens/assets-path.ts
@@ -1,0 +1,3 @@
+import {tuiCreateToken} from '@taiga-ui/cdk';
+
+export const TUI_ASSETS_PATH = tuiCreateToken('/assets/taiga-ui/icons/');
