+++ projects/core/tokens/index.ts
@@ -1,15 +1,14 @@
 export * from './animations-speed';
+export * from './assets-path';
 export * from './common-icon';
 export * from './date-format';
 export * from './day-type-handler';
 export * from './first-day-of-week';
 export * from './i18n';
 export * from './icon-resolver';
-export * from './icons';
 export * from './media';
 export * from './number-format';
 export * from './reduced-motion';
-export * from './sanitizer';
 export * from './scroll-ref';
 export * from './selection-stream';
 export * from './spin-icons';
