+++ projects/core/utils/dom/index.ts
@@ -1,4 +1,3 @@
 export * from './check-fixed-position';
 export * from './get-viewport';
 export * from './get-word-range';
-export * from './process-icon';
