+++ projects/core/utils/miscellaneous/index.ts
@@ -1,7 +1,5 @@
-export * from './icons-path-factory';
 export * from './is-editing-key';
 export * from './is-obscured';
-export * from './is-presumed-html-string';
 export * from './override-options';
 export * from './size-bigger';
 export * from './to-animation-options';
