+++ projects/core/utils/miscellaneous/test/is-presumed-html-string.spec.ts
@@ -1,4 +1,4 @@
-import {tuiIsPresumedHTMLString} from '../is-presumed-html-string';
+import {tuiIsPresumedHTMLString} from '@taiga-ui/legacy/utils/is-presumed-html-string';
 
 describe('isPresumedHTMLString', () => {
     describe('returns false if', () => {
