+++ projects/demo/src/modules/app/demo-routes.ts
@@ -132,9 +132,6 @@ export const DemoRoute = {
     Switch: '/components/switch',
     Tooltip: '/components/tooltip',
     Tree: '/components/tree',
-    IconsOverview: '/icons/overview',
-    IconsMapping: '/icons/mapping',
-    IconsBundled: '/icons/bundled',
     IconsCustomization: '/icons/customization',
     Filter: '/components/filter',
     ArcChart: '/charts/arc-chart',
