+++ projects/demo/src/modules/components/abstract/textfield-controller-documentation/index.html
@@ -30,15 +30,7 @@
         [documentationPropertyValues]="documentedComponent.customContentVariants"
         [(documentationPropertyValue)]="documentedComponent.customContentSelected"
     >
-        Right content (e.g. avatar with maximum size 32x32px). If it gets a string, it is inserted as
-        <code>src</code>
-        or as icon name into
-        <a
-            tuiLink
-            [routerLink]="routes.IconsOverview"
-        >
-            tui-svg
-        </a>
+        Right content (e.g. avatar with maximum size 32x32px)
     </ng-template>
     <ng-template
         documentationPropertyMode="input"
