+++ projects/demo/src/modules/components/icons/icons-bundled/examples/1/index.ts
@@ -5,9 +5,9 @@ import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {assets} from '@demo/utils';
 import {TUI_TRUE_HANDLER} from '@taiga-ui/cdk';
-import {TUI_SANITIZER, TuiSvgComponent, TuiSvgService} from '@taiga-ui/core';
 import {NgDompurifySanitizer} from '@taiga-ui/dompurify';
 import {tuiIconClockLarge, tuiIconMaestro, tuiIconMastercard} from '@taiga-ui/icons';
+import {TUI_SANITIZER, TuiSvgComponent, TuiSvgService} from '@taiga-ui/legacy';
 import {map, timer} from 'rxjs';
 
 @Component({
