--- projects/demo/src/modules/components/icons/icons-mapping/index.less
@@ -1,10 +1,0 @@
-.wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-
-.sanitizer-example {
-    display: block;
-    padding-top: 3.5rem;
-}
