+++ projects/demo/src/modules/directives/textfield-controller/index.html
@@ -71,15 +71,7 @@
                 [documentationPropertyValues]="customContentVariants"
                 [(documentationPropertyValue)]="customContentSelected"
             >
-                Right content (e.g. avatar with maximum size 32x32px). If it gets a string, it is inserted as
-                <code>src</code>
-                or as icon name into
-                <a
-                    tuiLink
-                    [routerLink]="routes.IconsOverview"
-                >
-                    tui-svg
-                </a>
+                Right content (e.g. avatar with maximum size 32x32px)
             </ng-template>
             <ng-template
                 documentationPropertyMode="input"
