+++ projects/legacy/components/svg/content-processor.ts
