+++ projects/legacy/components/svg/deprecated-icons.ts
