+++ projects/legacy/components/svg/index.ts
@@ -1,4 +1,5 @@
 export * from './content-processor';
 export * from './deprecated-icons';
 export * from './svg.component';
+export * from './svg.service';
 export * from './svg-options';
