+++ projects/legacy/components/svg/ng-package.json
