+++ projects/legacy/components/svg/svg-options.ts
@@ -2,7 +2,7 @@ import type {FactoryProvider, Provider} from '@angular/core';
 import {InjectionToken, Optional, SkipSelf} from '@angular/core';
 import type {TuiHandler, TuiSafeHtml, TuiStringHandler} from '@taiga-ui/cdk';
 import {tuiCreateTokenFromFactory, tuiIsString} from '@taiga-ui/cdk';
-import {tuiIconsPathFactory} from '@taiga-ui/core/utils';
+import {tuiIconsPathFactory} from '@taiga-ui/legacy/utils';
 import {identity} from 'rxjs';
 
 import {tuiSvgLinearGradientProcessor} from './content-processor';
