+++ projects/legacy/components/svg/svg.style.less
