+++ projects/legacy/components/svg/svg.template.html
