+++ projects/legacy/components/svg/test/svg-options.component.spec.ts
