+++ projects/legacy/components/svg/test/svg.component.spec.ts
