+++ projects/legacy/tokens/index.ts
@@ -1,11 +1,13 @@
 export * from './countries-masks';
 export * from './focusable-item-accessor';
 export * from './fonts-ready';
+export * from './icons';
 export * from './is-apple';
 export * from './is-chromium';
 export * from './is-firefox';
 export * from './is-stackblitz';
 export * from './month-formatter';
+export * from './sanitizer';
 export * from './textfield-appearance';
 export * from './textfield-host';
 export * from './touch-supported';
