+++ projects/legacy/utils/icons-path-factory.ts
