+++ projects/legacy/utils/is-presumed-html-string.ts
