+++ projects/legacy/utils/process-icon.ts
